refactor: removed Protocol as base class
