+++ app/factory/service.py
@@ -1,8 +1,5 @@
-import typing as t
-
 from flask import current_app
 
-from app.interfaces.factory.service import IServiceFactory
 from app.services.account import AccountService
 from app.services.authentication import AuthenticationService
 from app.services.export import ExportService
@@ -11,36 +8,27 @@ from app.services.report import ReportService
 from app.services.search import SearchService
 from app.services.signing import SigningService
 
-if t.TYPE_CHECKING:
-    from app.interfaces.services.account import IAccountService
-    from app.interfaces.services.authentication import IAuthenticationService
-    from app.interfaces.services.export import IExportService
-    from app.interfaces.services.mail import IMailService
-    from app.interfaces.services.report import IReportService
-    from app.interfaces.services.search import ISearchService
-    from app.interfaces.services.signing import ISigningService
 
-
-class ServiceFactory(IServiceFactory):
-    def get_export_service(self) -> "IExportService":
+class ServiceFactory:
+    def get_export_service(self):
         return ExportService()
 
-    def get_mail_service(self) -> "IMailService":
+    def get_mail_service(self):
         return MailService()
 
-    def get_report_service(self) -> "IReportService":
+    def get_report_service(self):
         return ReportService()
 
-    def get_search_service(self) -> "ISearchService":
+    def get_search_service(self):
         return SearchService()
 
-    def get_authentication_service(self) -> "IAuthenticationService":
+    def get_authentication_service(self):
         return AuthenticationService()
 
-    def get_account_service(self) -> "IAccountService":
+    def get_account_service(self):
         return AccountService()
 
-    def get_signing_service(self) -> "ISigningService":
+    def get_signing_service(self):
         return SigningService(
             current_app.config["SECRET_KEY"],
             current_app.config["SECURITY_PASSWORD_SALT"],
