+++ app/services/search.py
@@ -5,14 +5,13 @@ import sqlalchemy as sa
 
 from app.extensions import db
 from app.factories import dao_factory
-from app.interfaces.services.search import ISearchService
 from app.models.orm.lego import GenericSet, Theme, Year
 
 if t.TYPE_CHECKING:
     from sqlalchemy.orm import Session
 
 
-class SearchService(ISearchService):
+class SearchService:
     def __init__(self) -> None:
         self.session: "Session" = db.session
         self.theme_dao = dao_factory.get_theme_dao()
